Lab2/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nurs <nursultanzhantuar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"

[tool.project.scripts]
dev = 'src/manage:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
